- Updated Jetty to version 9.4.44.v20210927 - Updated Jackson to 2.13.0 - Updated Kafka client to 2.8.1 - Updated Vaadin to version 8.14.2 - Miscellaneous stability improvements

+++ src/main/java/org/point85/domain/jms/JmsClient.java
@@ -93,7 +93,7 @@ public class JmsClient extends BaseMessagingClient {
 
 		// non-transacted, auto-ack session
 		session = connection.createSession(false, Session.AUTO_ACKNOWLEDGE);
-		
+
 		setHostName(brokerHostName);
 		setHostPort(port);
 
@@ -285,7 +285,7 @@ public class JmsClient extends BaseMessagingClient {
 		textMessage.setText(serialize(message));
 		textMessage.setJMSCorrelationID(createCorrelationId());
 
-		producer.send(textMessage, DeliveryMode.NON_PERSISTENT, HIGH_PRIORITY, (long) (ttlSec * 1000));
+		producer.send(textMessage, DeliveryMode.NON_PERSISTENT, HIGH_PRIORITY, ttlSec * 1000l);
 
 		if (logger.isInfoEnabled()) {
 			logger.info("Sent text message of type " + message.getMessageType());
